EndianReader updates
No longer throws exceptions using the 'throw' keyword (which apparently has been depreciated anyway)
Exceptions are now handled internally, and can be used to gracefully recover, or even use them to gather more information about the state of the stream its self (Explained below).
Replaced all instances of const string ref with a string_view to mitigate unneeded copying.
Added a few more helper functions
isInBounds - Verifies a requested offset is in the bounds of the file. Sets the exception status to EXCEPTION_FILE_BOUNDS if exceeded and returns true or false
isOpen - Was a method before; however, now it sets the exception status to EXCEPTION_FILE_ACCESS as well as returning if it is open or not.
hasException - Tells if there has been an exception set
getException  - Returns a string_view containing the exception text.
clearException - Clears the exception (weird..)
releaseException - Returns a string view containing the exception text, and clears the exception at the same time.
isOpen - sets the exception status to EXCEPTION_FILE_ACCESS as well as returning if it is open or not.
Added functions that should of been there to start with
setFileEndianness - To modify a file's endianness. this may be needed if the stream is reused for a different file, or in the case of a mixed endian file.
readInto - A modification of the read function that reads a value into an existing memory location instead of creating a copy.
Added sanity checks
Set a maximum string recursion to prevent the stream from getting stuck in an infinite loop. ( currently set to 0xffffffff; however, may set to macro definition )
Seek, Read, Pad, and Peek all validate their offsets and sizes against the file size gracefully.
Read now assumes if offset in bounds of file, but the size of the data exceeds the file, the desired result is to read till end of file.  (EXCEPTION_FILE_BOUNDS will be set to let you know the chunk size is different than requested)
Updated the documentation for it one more time.